rename PromiseCanceller.ts file
